chore: Rename `safe-kit` to `sdk-starter-kit` (#939)
