+++ .github/workflows/sdk-test.yml
@@ -34,8 +34,8 @@ jobs:
         uses: coverallsapp/github-action@v2
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
-          files: packages/safe-kit/coverage/lcov.info
-            packages/api-kit/coverage/lcov.info
+          files: packages/api-kit/coverage/lcov.info
             packages/auth-kit/coverage/lcov.info
             packages/onramp-kit/coverage/lcov.info
             packages/relay-kit/coverage/lcov.info
+            packages/sdk-starter-kit/coverage/lcov.info
