+++ packages/sdk-starter-kit/package.json
@@ -1,18 +1,20 @@
 {
-  "name": "@safe-global/safe-kit",
+  "name": "@safe-global/sdk-starter-kit",
   "version": "1.0.0",
-  "description": "Safe Kit is a library that provides a set of tools to interact with the Safe Protocol.",
+  "description": "SDK that provides the basic tools to interact with the Safe Smart Account.",
   "main": "dist/src/index.js",
   "typings": "dist/src/index.d.ts",
   "keywords": [
-    "Safe",
     "Ethereum",
+    "Wallet",
+    "Safe",
+    "Smart Account",
     "Account Abstraction",
     "SDK"
   ],
   "scripts": {
     "unbuild": "rimraf dist",
-    "build": "yarn unbuild && tsc && tsc-alias",
+    "build": "yarn unbuild && tsc -p tsconfig.build.json && tsc-alias -p tsconfig.build.json",
     "format:check": "prettier --check \"*/**/*.{js,json,md,ts}\"",
     "format": "prettier --write \"*/**/*.{js,json,md,ts}\"",
     "test": "jest src --coverage"
