+++ packages/sdk-starter-kit/src/constants.ts
