+++ packages/sdk-starter-kit/src/extensions/messages/SafeMessageClient.test.ts
