+++ packages/sdk-starter-kit/src/extensions/safe-operations/types.ts
