+++ playground/sdk-starter-kit/send-off-chain-message.ts
@@ -1,4 +1,4 @@
-import { SafeClientResult, createSafeClient, offChainMessages } from '@safe-global/safe-kit'
+import { SafeClientResult, createSafeClient, offChainMessages } from '@safe-global/sdk-starter-kit'
 
 const OWNER_1_PRIVATE_KEY = ''
 const OWNER_2_PRIVATE_KEY = ''
