+++ tsconfig.json
@@ -1,12 +1,12 @@
 {
   "files": [],
   "references": [
-    { "path": "packages/safe-kit" },
     { "path": "packages/api-kit" },
     { "path": "packages/protocol-kit" },
     { "path": "packages/relay-kit" },
     { "path": "packages/auth-kit" },
     { "path": "packages/onramp-kit" },
-    { "path": "packages/safe-core-sdk-types" }
+    { "path": "packages/safe-core-sdk-types" },
+    { "path": "packages/sdk-starter-kit" }
   ]
 }
